+++ core/TreeNode.ts
@@ -2846,7 +2846,7 @@ class TreeNode extends AbstractNode {
     return str ? indent + str.replace(/\n/g, indent) : ""
   }
 
-  static getVersion = () => "49.6.1"
+  static getVersion = () => "49.7.0"
 
   static fromDisk(path: string): TreeNode {
     const format = this._getFileFormat(path)