+++ products/jtree.node.js
@@ -2150,6 +2150,10 @@ class TreeNode extends AbstractNode {
   has(firstWord) {
     return this._hasFirstWord(firstWord)
   }
+  hasNode(node) {
+    const needle = node.toString()
+    return this.getChildren().some(node => node.toString() === needle)
+  }
   _hasFirstWord(firstWord) {
     return this._getIndex()[firstWord] !== undefined
   }
@@ -3111,7 +3115,7 @@ TreeNode.iris = `sepal_length,sepal_width,petal_length,petal_width,species
 4.9,2.5,4.5,1.7,virginica
 5.1,3.5,1.4,0.2,setosa
 5,3.4,1.5,0.2,setosa`
-TreeNode.getVersion = () => "49.6.1"
+TreeNode.getVersion = () => "49.7.0"
 class AbstractExtendibleTreeNode extends TreeNode {
   _getFromExtended(firstWordPath) {
     const hit = this._getNodeFromExtended(firstWordPath)